Squelch signed vs. unsigned comparison warnings.
We know len is >= 1, hence > 0, at that point, so just cast it to
unsigned.

+++ bpf/net/bpf_filter.c
@@ -633,7 +633,7 @@ bpf_validate(f, len)
 		return 0;
 #endif
 
-	for (i = 0; i < len; ++i) {
+	for (i = 0; i < (u_int)len; ++i) {
 		p = &f[i];
 		switch (BPF_CLASS(p->code)) {
 		/*
